SwiftEval: Load tests bundle if exists.
Previously, when running tests through injection, if the test would have
depended upon a symbol that's in the tests bundle, it would give an
undefined symbol error. This commit solves this by loading the tests
bundle if it exists, which allows for much better support when loading
new test files through injection.